requirements.txt
asgiref==3.2.7
dj-database-url==0.5.0
Django==3.0.5
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
